pyproject.toml
[tool.poetry]
name = "ted_v2"
version = "0.0.1"
description = "Just another Discord bot written in Python...v2!"
authors = ["Alex The JPEG <falexthejpeg@gmail.com>"]
repository = "https://github.com/AlexTheJPEG/ted-v2"
documentation = "https://AlexTheJPEG.github.io/ted-v2/"
readme = "README.md"
packages = [
  {include = "ted_v2"}
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hikari = {extras = ["speedups"], version = "^2.0.0-dev.108"}
uvloop = "^0.16.0"
hikari-lightbulb = "^2.2.1"


[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.1"
mkdocs = "^1.3.0"
mkdocs-material = "^8.2.9"
mkdocstrings = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
target-version = ['py39']
fast = true
exclude = '''
(
  /(                        # exclude a few common directories in the
    \.git                   # root of the project
    | \.pytest_cache
    | python-venv
    | \.venv
    | build
    | dist
    | \.tox
  ))
'''

[tool.isort]
profile = "black"
